cursor/agents/requirements.txt
opencv-python
numpy
colorthief
Pillow
supabase
requests
